fix(subscription-addons): reflect overridden addon price in the Add Subscription table
The Charges column in SubscriptionAddonTable computed each addon's price from
a shared priceOverrides prop that SubscriptionForm never populates, so a
price overridden via "Override Price" in the addon editor still showed the
addon's original list price in the summary table (and on the resulting
subscription's Details/Edit pages before a reload). Each addon row now
builds its own overrides map from its own override_line_items, via a
overrideLineItemsToMap helper shared with SubscriptionAddonModal instead of
duplicated there.
